backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
redis[hiredis]==5.2.1
openai==1.58.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
sse-starlette==2.2.1
httpx==0.28.1
xmltodict==0.14.1
python-dotenv==1.0.1
pgvector==0.3.6
pypdf==5.1.0
python-docx==1.1.2
email-validator==2.1.1
